Fixed collision physics. Added predictive platform detection so that the object falling sets the correct height of the object when it collides with it to prevent the sprite from passing through the collider and getting stuck somewhere in between

+++ Checkers/Checkers/Collider.cpp
@@ -9,7 +9,7 @@ Collider::Collider()
 }
 
 
-void Collider::setBounds(GLfloat maxX, GLfloat maxY, GLfloat minX, GLfloat minY)
+void Collider::setBounds(GLuint maxX, GLuint maxY, GLuint minX, GLuint minY)
 {
 	this->maxX = maxX;
 	this->maxY = maxY;
